+++ lib/src/map/built_map.dart
@@ -228,3 +228,9 @@ class _BuiltMap<K, V> extends BuiltMap<K, V> {
 
   bool hasExactKeyAndValueTypes(Type key, Type value) => K == key && V == value;
 }
+
+/// Extensions for [BuiltMap] on [Map].
+extension BuiltMapExtension<K, V> on Map<K, V> {
+  /// Converts to a [BuiltMap].
+  BuiltMap<K, V> build() => BuiltMap<K, V>.of(this);
+}